Cleaned up some NMS reflection fragments
