+++ 1.19/src/main/java/me/doclic/noencryption/compatibility/CompatiblePacketListener.java
@@ -2,17 +2,32 @@ package me.doclic.noencryption.compatibility;
 
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.ChannelPromise;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.protocol.game.ClientboundPlayerChatPacket;
+import net.minecraft.util.Crypt.SaltSignaturePair;
 
-public interface CompatiblePacketListener {
+import java.util.Optional;
 
-    /**
-     * @return the packet that should be read
-     */
-    Object readPacket(ChannelHandlerContext channelHandlerContext, Object packet) throws Exception;
+public class CompatiblePacketListener {
 
-    /**
-     * @return the packet that should be written
-     */
-    Object writePacket(ChannelHandlerContext channelHandlerContext, Object packet, ChannelPromise promise) throws Exception;
+    public Object readPacket(ChannelHandlerContext channelHandlerContext, Object packet) throws Exception { return packet; }
+
+    public Object writePacket(ChannelHandlerContext channelHandlerContext, Object packet, ChannelPromise promise) throws Exception {
+        if (packet instanceof final ClientboundPlayerChatPacket clientboundPlayerChatPacket) {
+            final Optional<Component> unsignedContent = clientboundPlayerChatPacket.unsignedContent();
+
+            // recreate a new packet
+            return new ClientboundPlayerChatPacket(
+                    unsignedContent.orElse(clientboundPlayerChatPacket.signedContent()), // use unsigned content if available, this is the signed content field
+                    unsignedContent, // unsigned content field
+                    clientboundPlayerChatPacket.typeId(),
+                    clientboundPlayerChatPacket.sender(),
+                    clientboundPlayerChatPacket.timeStamp(),
+                    new SaltSignaturePair(0, new byte[0])); // salt signature field
+            }
+
+        return packet;
+
+    }
 
 }
