+++ 1.19/src/main/java/me/doclic/noencryption/compatibility/CompatiblePlayer.java
@@ -1,10 +1,15 @@
 package me.doclic.noencryption.compatibility;
 
 import io.netty.channel.Channel;
+import org.bukkit.craftbukkit.v1_19_R1.entity.CraftPlayer;
 import org.bukkit.entity.Player;
 
-public interface CompatiblePlayer {
+public class CompatiblePlayer {
 
-    Channel getChannel(Player player);
+    public Channel getChannel(Player player) {
+
+        return ((CraftPlayer) player).getHandle().connection.connection.channel;
+
+    }
 
 }
