+++ 1.19/src/main/resources/plugin.yml
@@ -1,5 +1,5 @@
 name: NoEncryption
-version: '${project.version}'
+version: '${project.parent.version}'
 main: me.doclic.noencryption.NoEncryption
 api-version: 1.19
 authors: [ Doclic ]
